fix: mai erros fixed

+++ FrontEnd/FormBuilderAngular/src/app/components/deleterComponents/form-group-deleter/form-group-deleter.component.ts
@@ -1,4 +1,5 @@
 import {
+  AfterViewInit,
   Component,
   ElementRef,
   EventEmitter,
@@ -23,20 +24,38 @@ export class FormGroupDeleterComponent {
 
   constructor(private formGroupService: FormGroupService) {}
 
-  deleteFormGroup(formGroup: FormGroupModel): void {
-    this.formGroupService.deleteFormGroup(formGroup.id).subscribe({
-      next: () => {
-        console.log('Successfully deleting form Group');
-      },
-      error: (err) => {
-        console.error('Error deleting form Group', err);
-      },
-    });
-  }
+  // ngAfterViewInit(): void {
+  //   if (!this.dialogDeleteMessage) {
+  //     console.error('Dialog element not found or not yet initialized');
+  //   }
+  // }
 
-  closeDialogMessageDelete(): void {
-    if (this.dialogDeleteMessage && this.dialogDeleteMessage.nativeElement) {
-      this.dialogDeleteMessage.nativeElement.close();
+  deleteFormGroup(formGroup: FormGroupModel): void {
+    const confirmMsg = confirm('Are you sure about this?');
+
+    if (confirmMsg) {
+      this.formGroupService.deleteFormGroup(formGroup.id).subscribe({
+        next: () => {
+          console.log('Successfully deleting form Group');
+        },
+        error: (err) => {
+          console.error('Error deleting form Group', err);
+        },
+      });
     }
   }
+
+  // closeDialogMessageDelete(): void {
+  //   if (this.dialogDeleteMessage && this.dialogDeleteMessage.nativeElement) {
+  //     this.dialogDeleteMessage.nativeElement.close();
+  //   }
+  // }
+
+  // openDialog(): void {
+  //   if (this.dialogDeleteMessage && this.dialogDeleteMessage.nativeElement) {
+  //     this.dialogDeleteMessage.nativeElement.showModal();
+  //   } else {
+  //     console.error('Dialog element not found or not yet initialized.');
+  //   }
+  // }
 }
